feat(project): project's endpoints and entity edited
This commit removes the `ProjectController` from `ApiModule` and adds new endpoints to `ProjectController`. The `GetProjectById` DTO is added to validate the id parameter in the getProjectById endpoint. The start and end dates are now validated using a custom decorator called IsDateString. A TypeORM module is imported into the project module, which allows for easy database integration. Finally, an HttpException is thrown when no content is found in the getProjectById method of ProjectService.