feat(projects): Add user entity relation to project entity
This commit adds a ManyToOne relationship between the User and Project entities. The createdBy field in the Project entity is now linked to the User entity. Additionally, the createProject method in the ProjectService now accepts a User object as a parameter and assigns it to the createdBy field of the new project.