fix(invoice): Fix generatePdf endpoint and add GeneratePdfDto
The commit fixes the `generate-pdf` endpoint in the `InvoiceController` by changing it from a GET request to a POST request. It also adds a new DTO called `GeneratePdfDto` which is used as the request body for the endpoint.

The changes include:
- Changed the HTTP method of the `generate-pdf` endpoint from GET to POST.
- Added a new DTO called `GeneratePdfDto`.
- Updated the parameters of the `downloadPdf` method in the `InvoiceController` to use the new DTO.
- Updated the parameters of the `generatePdf` method in the `InvoiceService` to use a specific type for project.
- Modified the PDF generation logic in the `generatePdf` method of the `InvoiceService`.

These changes ensure that proper data is sent to and received from the server when generating PDFs for invoices.